LC-640-fix Sprint7 1차 QA 이슈 해결

+++ src/main/java/org/letscareer/letscareer/domain/report/service/impl/DeleteReportServiceImpl.java
@@ -4,6 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.letscareer.letscareer.domain.report.entity.Report;
 import org.letscareer.letscareer.domain.report.helper.ReportHelper;
 import org.letscareer.letscareer.domain.report.service.DeleteReportService;
+import org.letscareer.letscareer.domain.user.entity.User;
+import org.letscareer.letscareer.domain.user.helper.UserHelper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,10 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 @Service
 public class DeleteReportServiceImpl implements DeleteReportService {
+    private final UserHelper userHelper;
     private final ReportHelper reportHelper;
 
     @Override
-    public void execute(Long reportId) {
+    public void execute(User user, Long reportId) {
+        userHelper.validateAdminUser(user);
         Report report = reportHelper.findReportByReportIdOrThrow(reportId);
         reportHelper.deleteReport(report);
     }
